api/requirements.txt
# FastAPI 코어
fastapi==0.115.12
uvicorn[standard]==0.34.3

# 데이터베이스
sqlalchemy==2.0.41
alembic==1.16.1
aiosqlite==0.21.0          # 개발용 SQLite (async)
asyncpg==0.30.0            # 프로덕션용 PostgreSQL (async)

# 인증 / 보안
python-jose[cryptography]==3.3.0   # JWT 생성/검증
passlib[bcrypt]==1.7.4             # 비밀번호 해싱
python-multipart==0.0.20           # Form data 처리

# 유효성 검사
pydantic==2.11.5
pydantic-settings==2.9.1
email-validator==2.2.0

# 캐시 (Redis - 선택적)
redis==5.2.1

# HTTP 클라이언트 (AI API 호출용)
httpx==0.28.1

# 유틸리티
python-dotenv==1.1.0
